refactor: FocusPage 컴포넌트 폴더 구조 분리 및 import 경로 수정
