+++ src/pages/FocusPage/Focus.jsx
@@ -4,13 +4,13 @@ import { getStudyDetail } from "../../api/studies";
 import Toast from "../../components/Toast/Toast";
 import useTimer from "../../hooks/timer/useTimer";
 import { TIMER_STATUS } from "../../hooks/timer/timerConstants";
-import TimerSetting from "./components/TimerSetting";
-import TimerControls from "./components/TimerControls";
-import ResumeConfirmPopup from "./components/ResumeConfirmPopup";
-import StopConfirmPopup from "./components/StopConfirmPopup";
-import SessionTitleModal from "./components/SessionTitleModal";
-import SessionListModal from "./components/SessionListModal";
-import StudyHeader from "./components/StudyHeader";
+import TimerSetting from "./components/timer/TimerSetting";
+import TimerControls from "./components/timer/TimerControls";
+import ResumeConfirmPopup from "./components/popups/ResumeConfirmPopup";
+import StopConfirmPopup from "./components/popups/StopConfirmPopup";
+import SessionTitleModal from "./components/session/SessionTitleModal";
+import SessionListModal from "./components/session/SessionListModal";
+import StudyHeader from "./components/header/StudyHeader";
 import formatTime from "./formatTime";
 import styles from "./Focus.module.css";
