+++ src/pages/FocusPage/components/session/SessionTitleModal.jsx
@@ -1,4 +1,4 @@
-import Modal from "../../../components/Modal/Modal";
+import Modal from "../../../../components/Modal/Modal";
 import styles from "./SessionTitleModal.module.css";
 
 function SessionTitleModal({
