+++ src/pages/FocusPage/components/timer/TimerSetting.module.css
